Refactor
There was a bug in adding a course with CRN, in which repetidely clicking on the add button with an empty field would cause the app to crash.
Fix: the app no langer crashes with the repetitive clicking of the add button regardless of whether something is in the field or not.

Ritobrata Sen
Qu Yuze